+++ NEWS.md
@@ -3,6 +3,7 @@ Red Panda C++ Version 2.8
   - fix: Crash when editing makefile
   - enhancement: Add "Resources" in project option's dialog's custom compiler parameter page
   - fix: Crash while input using input method in makefile 
+  - enhancement: "Run" / "Generate Assembly" for project source files
 
 Red Panda C++ Version 2.7